apps/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
cryptography==45.0.6
numpy==2.0.2
scikit-learn==1.5.2
